Lightbox with red cross
Made a very rudimentary 'x-out' icon for the lightbox. So now you don't
click on the image to exit the lightbox, you click on the X. Probably
Ryan could mock up an art asset better looking than the thing I've done,
but this means now that the user can interact with the image in the
lightbox without exiting out of it, which means we can put whatever kind
of interactivity we want on top of it.